generate slides based on output from enableRF

generate slides based on output from enableRF

generate_slides(
  outputs,
  outfile = "output.pptx",
  template = file.path(system.file(package = "autoslider.core"), "theme/basic.pptx"),
  fig_width = 9,
  fig_height = 6,
  t_lpp = 20,
  t_cpp = 200,
  l_lpp = 20,
  l_cpp = 150,
  ...
)

Arguments

outputs

List of output

outfile

Out file path

template

Template file path

fig_width

figure width in inch

fig_height

figure height in inch

t_lpp

An integer specifying the table lines per page
Specify this optional argument to modify the length of all of the table displays

t_cpp

An integer specifying the table columns per page
Specify this optional argument to modify the width of all of the table displays

l_lpp

An integer specifying the listing lines per page
Specify this optional argument to modify the length of all of the listings display

l_cpp

An integer specifying the listing columns per page
Specify this optional argument to modify the width of all of the listings display

...

arguments passed to program

Examples

# Example 1. When applying to the whole pipeline
data <- list(
  adsl = eg_adsl %>% dplyr::mutate(FASFL = SAFFL),
  adae = eg_adae
)

filters::load_filters(
  yaml_file = system.file("filters.yml", package = "autoslider.core"),
  overwrite = TRUE
)

spec_file <- system.file("spec.yml", package = "autoslider.core")
spec_file %>%
  read_spec() %>%
  filter_spec(program %in% c("t_dm_slide")) %>% # removed "l_dsl01_slide" atm
  generate_outputs(datasets = data) %>%
  decorate_outputs() %>%
  generate_slides()
#> Error in get_version_label_output(): could not find function "get_version_label_output"

# Example 2. When applying to an rtable object or an rlisting object
adsl <- eg_adsl
t_dm_slide(adsl, "TRT01P", c("SEX", "AGE")) %>%
  generate_slides()
#> Error in t_dm_slide_to_df(result): could not find function "t_dm_slide_to_df"

l_dsl01_slide(adsl) %>% generate_slides()
#> Error in l_dsl01_slide(adsl): could not find function "l_dsl01_slide"