Read yaml spec file

Read yaml spec file and split according to filter lists

read_spec(spec_file = "spec.yml", metadata = NULL)

Arguments

spec_file

`character`. Path to a yaml spec file

metadata

Metadata of study

Value

An object of class `spec` which is a `list` where each element corresponds to one output, e.g. `t_dm_IT`.

Author

- Liming Li (`Lil128`) - Thomas Neitmann (`neitmant`)

Examples

spec_file <- system.file("spec.yml", package = "autoslideR")

## Take a look at the 'raw' content of the spec file
cat(readLines(spec_file)[1:24], sep = "\n")
#> Warning: file("") only supports open = "w+" and open = "w+b": using the former
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA
#> NA

## This is how it looks once read into R
spec <- read_spec(spec_file)
#> Warning: file("") only supports open = "w+" and open = "w+b": using the former
spec[1:3]
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#>